requirements.txt
Requests
Pillow 
Flask